refactor: 重构任务详情展示方式，优化附件处理
1. 将任务详情页改为弹窗形式
- 删除 TaskDetail 页面
- 新增 TaskDetailModal 组件
- 修改 TaskCenter 和 MyTasks 使用弹窗
- 确保子弹窗层级正确

2. 优化任务附件处理
- 规范化附件存储路径和文件名
- 添加任务ID和时间戳到文件名
- 使用UUID确保文件名唯一

3. 修复任务放弃功能
- 完善数据清理逻辑
- 确保附件文件被正确删除
- 清空相关时间戳和说明信息
